Re-organize examples to improve grouping and readability (#817)
