+++ doc/changelog.d/817.documentation.md
@@ -1,0 +1,1 @@
+Re-organize examples to improve grouping and readability
