--- examples/1_Impacted_Substances_Queries/index.rst
@@ -1,16 +1,0 @@
-.. _ref_grantami_bomanalytics_impactedsubstances:
-
-Impacted substances examples
-============================
-
-These examples show how to perform an impacted substance query on material records or parts in a Granta MI Restricted
-Substances database, identifying any substances impacted by one or more legislations. You can write an equivalent script
-for specifications. For more information, see :ref:`ref_grantami_bomanalytics_api_index`. To determine the substances impacted
-by an XML BoM, see :ref:`ref_grantami_bomanalytics_advanced`.
-
-
-.. toctree::
-   :maxdepth: 1
-
-   1-1_Materials_impacted_substances.ipynb
-   1-2_Parts_impacted_substances.ipynb
