--- examples/2_Compliance_Queries/index.rst
@@ -1,17 +1,0 @@
-.. _ref_grantami_bomanalytics_compliance:
-
-Compliance Examples
-===================
-
-These examples show how to determine compliance against one or more legislations for substances,
-materials, or parts in a Granta MI Restricted Substances database. An equivalent script can be written for
-specifications. For more information, see :ref:`ref_grantami_bomanalytics_api_index`.
-
-To determine compliance of an XML BoM, see :ref:`ref_grantami_bomanalytics_advanced`.
-
-.. toctree::
-   :maxdepth: 1
-
-   2-1_Substance_compliance.ipynb
-   2-2_Material_compliance.ipynb
-   2-3_Part_compliance.ipynb
