+++ examples/2_Impacted_Substances_Queries/2-2_Parts_impacted_substances.py
@@ -27,7 +27,7 @@
 # ## Connect to Granta MI
 
 # Import the ``Connection`` class and create the connection. For more information, see the
-# [Getting Started](../0_Getting_started.ipynb) example.
+# [Basic Usage](../0_Basic_usage.ipynb) example.
 
 # + tags=[]
 from ansys.grantami.bomanalytics import Connection
