+++ examples/2_Impacted_Substances_Queries/index.rst
@@ -1,0 +1,16 @@
+.. _ref_grantami_bomanalytics_impactedsubstances:
+
+Impacted substances examples
+============================
+
+These examples show how to perform an impacted substance query on materials, parts, or XML BoMs in a Granta MI
+Restricted Substances database, identifying any substances impacted by one or more legislations. You can write an
+equivalent script for specifications. For more information, see :ref:`ref_grantami_bomanalytics_api_index`.
+
+
+.. toctree::
+   :maxdepth: 1
+
+   2-1_Materials_impacted_substances
+   2-2_Parts_impacted_substances
+   2-3_BoM_impacted_substances
