--- examples/3_Advanced_Topics/index.rst
@@ -1,16 +1,0 @@
-.. _ref_grantami_bomanalytics_advanced:
-
-Advanced examples
-=================
-
-These examples demonstrate various advanced functionalities of Granta MI BoM Analytics.
-
-.. toctree::
-   :maxdepth: 1
-
-   3-1_Working_with_XML_BoMs.ipynb
-   3-2_Creating_an_XML_BoM.ipynb
-   3-3_Dealing_with_external_data_sources.ipynb
-   3-4_Database-specific_configuration.ipynb
-   3-5_Writing_compliance_results_to_a_dataframe.ipynb
-   3-6_Creating_a_Bill_of_Materials.ipynb
