+++ examples/3_Compliance_Queries/3-1_Substance_compliance.py
@@ -15,14 +15,14 @@
 
 # # Perform a substance compliance query
 
-# A substance cmpliance query determines whether one or more substances are compliant with the specified indicators.
+# A substance compliance query determines whether one or more substances are compliant with the specified indicators.
 # This example checks several materials for substances included on two watch lists ("EU REACH - The Candidate List" and
 # "The SIN List 2.1"), specifying substance amounts and thresholds for compliance.
 
 # ## Connect to Granta MI
 
 # Import the ``Connection`` class and create the connection. For more information, see the
-# [Getting Started](../0_Getting_started.ipynb) example.
+# [Basic Usage](../0_Basic_usage.ipynb) example.
 
 # + tags=[]
 from ansys.grantami.bomanalytics import Connection
