+++ examples/3_Compliance_Queries/3-4_BoM_compliance.py
@@ -13,14 +13,13 @@
 #     name: python3
 # ---
 
-# # Perform a BoM query
+# # Perform a BoM compliance query
 
-# Both impacted substances queries and compliance queries can be performed on an XML BoM instead of a list
-# of Granta MI records. The BoM must be in the Granta 17/11 format or Granta 23/01 format. This example shows how to use
+# Compliance queries can be performed on an XML BoM instead of a list
+# of Granta MI records. The BoM must be in any valid Granta XML BoM format. This example shows how to use
 # the ``lxml`` package with the XSD XML schema file to validate the XML format.
 
-# For help on constructing an XML BoM, see the
-# [Creating an XML Bill of Materials](3-2_Creating_an_XML_BoM.py) example.
+# For help on constructing an XML BoM, see [BoM examples](../6_BoMs/index.rst).
 
 # If the XML file is generated by a Granta MI product and has not been modified, it is possible to skip this step before
 # submitting the query. However, it is strongly advised to validate the XML BoM in all situations to avoid unexpected
@@ -30,8 +29,8 @@
 
 # The following supporting files are required for this example:
 #
-# * [compliance-bom.xml](supporting-files/compliance-bom.xml)
-# * [invalid-bom.xml](supporting-files/invalid-bom.xml)
+# * [compliance-bom.xml](../supporting-files/compliance-bom.xml)
+# * [invalid-bom.xml](../supporting-files/invalid-bom.xml)
 
 # The XSD XML schema is included in the library in the ``schemas`` subpackage.
 
@@ -49,8 +48,6 @@ def xml_validator(xml: str, schema_file: str) -> bool:
     schema = etree.XMLSchema(file=schema_file)
     doc = etree.fromstring(xml)
     return schema.validate(doc)
-
-
 # -
 
 # You can now use this function to test the validity of a BoM generated by BoM Analyzer.
@@ -58,7 +55,7 @@ def xml_validator(xml: str, schema_file: str) -> bool:
 # +
 from ansys.grantami.bomanalytics.schemas import bom_schema_1711
 
-valid_xml_file = "supporting-files/compliance-bom.xml"
+valid_xml_file = "../supporting-files/compliance-bom.xml"
 with open(valid_xml_file) as f:
     valid_xml = f.read()
 result = xml_validator(valid_xml, bom_schema_1711)
@@ -67,13 +64,13 @@ result
 
 # You can also test a BoM that is valid XML but is not compliant with the schema.
 
-invalid_xml_file = "supporting-files/invalid-bom.xml"
+invalid_xml_file = "../supporting-files/invalid-bom.xml"
 with open(invalid_xml_file) as f:
     invalid_xml = f.read()
 result = xml_validator(invalid_xml, bom_schema_1711)
 result
 
-# ## Run an impacted substances XML-based query
+# ## Run a compliance XML-based query
 
 # Now that you have validated the XML, you can build your XML BoM-based query.
 # First, connect to Granta MI.
@@ -85,50 +82,16 @@ server_url = "http://my_grantami_server/mi_servicelayer"
 cxn = Connection(server_url).with_credentials("user_name", "password").connect()
 # -
 
-# The impacted substances BoM query behaves similar to other impacted substances queries. However,
-# a BoM query can only accept a single BoM at a time, and so you only ever receive a single list
-# of substances impacted by the specified legislations.
-
-# +
-from ansys.grantami.bomanalytics import queries
-
-SIN_LIST = "SINList"
-impacted_substances_query = (
-    queries.BomImpactedSubstancesQuery()
-    .with_bom(valid_xml)
-    .with_legislation_ids([SIN_LIST])
-)
-# -
-
-impacted_substances_result = cxn.run(impacted_substances_query)
-impacted_substances_result
-
-# The ``BomImpactedSubstancesQueryResult`` object returned after running the query for impacted substances now behaves
-# similarly to the result object for any other query for impacted substances. For example, you can print all substances
-# impacted by the legislation using an approach from the previous example.
-
-# +
-from tabulate import tabulate
-
-rows = [
-    [substance.cas_number, substance.max_percentage_amount_in_material]
-    for substance in impacted_substances_result.impacted_substances
-]
-print(f'Substances impacted by "{SIN_LIST}" in a BoM (5/{len(rows)})')
-print(tabulate(rows[:5], headers=["CAS Number", "Amount (wt. %)"]))
-# -
-
-# ## Run a compliance XML-based query
-
 # Running a BoM compliance query produces the same result as if you had stored the BoM structure as linked **Products
 # and Parts** records in Granta MI.
 
 # +
-from ansys.grantami.bomanalytics import indicators
+from ansys.grantami.bomanalytics import indicators, queries
 
+CANDIDATE_LIST = "Candidate_AnnexXV"
 svhc = indicators.WatchListIndicator(
     name="SVHC",
-    legislation_ids=["Candidate_AnnexXV"],
+    legislation_ids=[CANDIDATE_LIST],
     default_threshold_percentage=0.1,
 )
 compliance_query = (
@@ -158,7 +121,7 @@ print(f"BoM Compliance Status: {root_part.indicators['SVHC'].flag.name}")
 broken_query = (
     queries.BomImpactedSubstancesQuery()
     .with_bom(invalid_xml)
-    .with_legislation_ids([SIN_LIST])
+    .with_legislation_ids([CANDIDATE_LIST])
 )
 
 RUN_QUERY = False
