+++ examples/3_Compliance_Queries/index.rst
@@ -1,0 +1,18 @@
+.. _ref_grantami_bomanalytics_compliance:
+
+Compliance examples
+===================
+
+These examples show how to determine compliance against one or more legislations for substances,
+materials, parts, or XML BoMs in a Granta MI Restricted Substances database. An equivalent script can be written for
+specifications. For more information, see :ref:`ref_grantami_bomanalytics_api_index`.
+
+.. toctree::
+   :maxdepth: 1
+
+   3-1_Substance_compliance
+   3-2_Material_compliance
+   3-3_Part_compliance
+   3-4_BoM_compliance
+   3-5_Compliance_for_external_data
+   3-6_Writing_compliance_results_to_a_dataframe
