+++ examples/4_Sustainability/index.rst
@@ -1,13 +1,12 @@
 .. _ref_grantami_bomanalytics_sustainability:
 
-Sustainability
-==============
+Sustainability examples
+=======================
 
-These examples show how to determine the environmental impact for a BoM using the
+The example shows how to determine the environmental impact for a BoM using the
 :ref:`ref_grantami_bomanalytics_api_sustainability_index`.
 
 .. toctree::
    :maxdepth: 1
 
-   4-1_BoM_Sustainability_summary.ipynb
-   4-2_BoM_Sustainability.ipynb
+   4-1_Sustainability
