+++ examples/5_Sustainability_summary/index.rst
@@ -1,0 +1,42 @@
+.. _ref_grantami_bomanalytics_sustainability_bom_sustainability_summary:
+
+Sustainability summary examples
+===============================
+
+The examples in this section demonstrate different aspects of the :class:`.BomSustainabilitySummaryQueryResult` object:
+
+* :doc:`5-1_Summary_and_messages`
+
+  * :attr:`~.BomSustainabilitySummaryQueryResult.messages`
+  * :attr:`~.BomSustainabilitySummaryQueryResult.phases_summary`
+
+* :doc:`5-2_Transports`
+
+  * :attr:`~.BomSustainabilitySummaryQueryResult.transport_details`
+  * :attr:`~.BomSustainabilitySummaryQueryResult.distribution_transport_summary`
+  * :attr:`~.BomSustainabilitySummaryQueryResult.manufacturing_transport_summary`
+  * :attr:`~.BomSustainabilitySummaryQueryResult.transport_details_aggregated_by_part`
+
+* :doc:`5-3_Materials`
+
+  * :attr:`~.BomSustainabilitySummaryQueryResult.material_details`
+
+* :doc:`5-4_Processes`
+
+  * :attr:`~.BomSustainabilitySummaryQueryResult.primary_processes_details`
+  * :attr:`~.BomSustainabilitySummaryQueryResult.secondary_processes_details`
+  * :attr:`~.BomSustainabilitySummaryQueryResult.joining_and_finishing_processes_details`
+
+* :doc:`5-5_Hierarchical_plots`
+
+  * Summarizes all the processed data into plots showing the result hierarchy.
+
+.. toctree::
+   :maxdepth: 1
+   :hidden:
+
+   5-1_Summary_and_messages
+   5-2_Transports
+   5-3_Materials
+   5-4_Processes
+   5-5_Hierarchical_plots
