+++ examples/6_BoMs/6-1_Creating_an_XML_BoM.py
