+++ examples/6_BoMs/index.rst
@@ -1,0 +1,12 @@
+.. _ref_grantami_bomanalytics_advanced:
+
+BoM examples
+============
+
+These examples demonstrate how to create XML BoMs.
+
+.. toctree::
+   :maxdepth: 1
+
+   6-1_Creating_an_XML_BoM
+   6-2_Creating_an_XML_BoM_from_JSON
