+++ examples/index.rst
@@ -11,13 +11,67 @@ To run these examples, install dependencies with this code:
 
    pip install ansys-grantami-bomanalytics[examples]
 
+.. grid:: 1 2 2 2
+
+   .. grid-item-card:: Basic usage example
+      :padding: 2 2 2 2
+      :link: 0_Basic_usage
+      :link-type: doc
+
+      How to connect to Granta MI and perform a basic query
+
+   .. grid-item-card:: Database-specific configuration
+      :padding: 2 2 2 2
+      :link: 1_Database-specific_configuration
+      :link-type: doc
+
+      How to configure BoM Analytics for a customized database
+
+   .. grid-item-card:: Impacted substances examples
+      :padding: 2 2 2 2
+      :link: 2_Impacted_Substances_Queries/index
+      :link-type: doc
+
+      How to perform impacted substance queries for materials, parts, and BoMs
+
+   .. grid-item-card:: Compliance examples
+      :padding: 2 2 2 2
+      :link: 3_Compliance_Queries/index
+      :link-type: doc
+
+      How to perform compliance substance queries for substances, materials, parts,
+      and with external data sources
+
+   .. grid-item-card:: Sustainability examples
+      :padding: 2 2 2 2
+      :link: 4_Sustainability/index
+      :link-type: doc
+
+      How to perform sustainability queries and to analyze the results
+
+   .. grid-item-card:: Sustainability summary examples
+      :padding: 2 2 2 2
+      :link: 5_Sustainability_summary/index
+      :link-type: doc
+
+      How to perform sustainability summary queries and to analyze the results
+
+   .. grid-item-card:: BoM examples
+      :padding: 2 2 2 2
+      :link: 6_BoMs/index
+      :link-type: doc
+
+      How to create XML BoMs
+
 
 .. toctree::
    :maxdepth: 2
+   :hidden:
 
-   0_Getting_started
-   1_Impacted_Substances_Queries/index
-   2_Compliance_Queries/index
-   3_Advanced_Topics/index
+   0_Basic_usage
+   1_Database-specific_configuration
+   2_Impacted_Substances_Queries/index
+   3_Compliance_Queries/index
    4_Sustainability/index
-
+   5_Sustainability_summary/index
+   6_BoMs/index
