+++ examples/supporting-files/invalid-bom.xml
