+++ examples/supporting-files/sustainability-bom-2412.xml
